backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
google-cloud-vision==3.7.4
anthropic==0.40.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
